remove pipdeps, update Makefile
