+++ acceptancetests/Makefile
@@ -4,8 +4,8 @@ test:
 	TMPDIR=/tmp python -m unittest discover -vv . -p "$(p)"
 	TMPDIR=/tmp python3 -m unittest discover -vv jujupy -t . -p "$(p)"
 lint:
-	python3 -m flake8 --builtins xrange,basestring $(py3)
-	flake8 $$(find -name '*.py') --builtins xrange,basestring --exclude $(py3)
+	python3 -m flake8 --builtins xrange,basestring $(py3) --exclude=repository
+	flake8 --builtins xrange,basestring --exclude=$(py3),repository
 cover:
 	python -m coverage run --source="./" --omit "./tests/*" -m unittest discover -vv ./tests
 	python -m coverage report
@@ -21,7 +21,7 @@ juju-ci-tools.common_0.1.4-0_all.deb: apt-update
 install-deps: juju-ci-tools.common_0.1.4-0_all.deb apt-update
 	sudo dpkg -i juju-ci-tools.common_0.1.4-0_all.deb || true
 	sudo apt-get install -y -f
-	./pipdeps.py -v install
+	pip install -r $HOME/juju/acceptancetests/requirements.txt
 name=NAMEHERE
 assess_file=assess_$(name).py
 new-assess:
