+++ acceptancetests/requirements.txt
@@ -1,30 +1,31 @@
 apache-libcloud>=2.2.1
-azure==2.0.0rc3
-azure-batch==0.30.0rc3
-azure-common==1.1.3
-azure-graphrbac==0.30.0rc3
-azure-mgmt==0.30.0rc3
-azure-mgmt-authorization==0.30.0rc3
-azure-mgmt-batch==0.30.0rc3
-azure-mgmt-cdn==0.30.0rc3
-azure-mgmt-compute==0.30.0rc3
-azure-mgmt-logic==0.30.0rc3
-azure-mgmt-network==0.30.0rc3
-azure-mgmt-notificationhubs==0.30.0rc3
-azure-mgmt-nspkg==1.0.0
-azure-mgmt-redis==0.30.0rc3
-azure-mgmt-resource==0.30.0rc3
-azure-mgmt-scheduler==0.30.0rc3
-azure-mgmt-storage==0.30.0rc3
-azure-mgmt-web==0.30.0rc3
-azure-nspkg==1.0.0
-azure-servicebus==0.20.1
-azure-servicemanagement-legacy==0.20.3
+azure>=2.0.0rc3
+azure-batch>=0.30.0rc3
+azure-common>=1.1.3
+azure-graphrbac>=0.30.0rc3
+azure-mgmt>=0.30.0rc3
+azure-mgmt-authorization>=0.30.0rc3
+azure-mgmt-batch>=0.30.0rc3
+azure-mgmt-cdn>=0.30.0rc3
+azure-mgmt-compute>=0.30.0rc3
+azure-mgmt-logic>=0.30.0rc3
+azure-mgmt-network>=0.30.0rc3
+azure-mgmt-notificationhubs>=0.30.0rc3
+azure-mgmt-nspkg>=1.0.0
+azure-mgmt-redis>=0.30.0rc3
+azure-mgmt-resource>=0.30.0rc3
+azure-mgmt-scheduler>=0.30.0rc3
+azure-mgmt-storage>=0.30.0rc3
+azure-mgmt-web>=0.30.0rc3
+azure-nspkg>=1.0.0
+azure-servicebus>=0.20.1
+azure-servicemanagement-legacy>=0.20.3
 azure-storage>=0.31.0
 boto>=2.48
 isodate>=0.5.1
 msgpack-python>=0.4.6
 msrestazure>=0.3.0
-pywinrm==0.0.3
+pywinrm>=0.0.3
 pexpect>=2.4
 xmltodict>=0.9.2
+python-novaclient>=6.0.2
